+++ patches/tModLoader/Terraria/ModLoader/ModCustomTree.cs
@@ -13,248 +13,6 @@ using Terraria.ObjectData;
 
 namespace Terraria.ModLoader;
 
-public static class CustomTreeLoader
-{
-	public const int TreeStyleBase = 1000;
-
-	internal static Dictionary<ushort, ModCustomTree> byTileLookup = new();
-	internal static Dictionary<ushort, ModCustomTree> bySaplingLookup = new();
-	internal static Dictionary<int, ModCustomTree> byAcornLookup = new();
-	internal static Dictionary<int, ModCustomTree> byCustomLeafLookup = new();
-
-	internal static List<ModCustomTree> trees = new();
-
-	internal static void Unload()
-	{
-		byTileLookup.Clear();
-		bySaplingLookup.Clear();
-		byAcornLookup.Clear();
-		byCustomLeafLookup.Clear();
-		trees.Clear();
-	}
-
-	internal static int NextTreeStyle()
-	{
-		return trees.Count + TreeStyleBase;
-	}
-
-	public static ModCustomTree GetByTile(ushort tile)
-	{
-		if (byTileLookup.TryGetValue(tile, out ModCustomTree tree))
-			return tree;
-		return null;
-	}
-    public static ModCustomTree GetBySapling(ushort sapling)
-	{
-		if (bySaplingLookup.TryGetValue(sapling, out ModCustomTree tree))
-			return tree;
-		return null;
-	}
-	public static ModCustomTree GetByAcorn(int acorn)
-	{
-		if (byAcornLookup.TryGetValue(acorn, out ModCustomTree tree))
-			return tree;
-		return null;
-	}
-	public static ModCustomTree GetByTreeStyle(int style)
-	{
-		if (style < TreeStyleBase)
-			return null;
-		style -= TreeStyleBase;
-		if (style >= trees.Count)
-			return null;
-
-		return trees[style];
-	}
-	public static ModCustomTree GetByCustomLeaf(int leaf)
-	{
-		if (byCustomLeafLookup.TryGetValue(leaf, out ModCustomTree tree))
-			return tree;
-		return null;
-	}
-
-	public static bool TryGetByTile(ushort tile, out ModCustomTree tree)
-	{
-		return byTileLookup.TryGetValue(tile, out tree);
-	}
-    public static bool TryGetBySapling(ushort sapling, out ModCustomTree tree)
-	{
-		return bySaplingLookup.TryGetValue(sapling, out tree);
-	}
-	public static bool TryGetByAcorn(int acorn, out ModCustomTree tree)
-	{
-		return byAcornLookup.TryGetValue(acorn, out tree);
-	}
-	public static bool TryGetByTreeStyle(int style, out ModCustomTree tree)
-	{
-		tree = null;
-		if (style < TreeStyleBase)
-			return false;
-		style -= TreeStyleBase;
-		if (style >= trees.Count)
-			return false;
-
-		tree = trees[style];
-		return true;
-	}
-	/// <summary>
-	/// Gets tree foliage data for rendering
-	/// </summary>
-	/// <param name="x">World X coordinate</param>
-	/// <param name="y">World Y coordinate</param>
-	/// <param name="xoffset"></param>
-	/// <param name="treeFrame">Tree frame (0, 1, 2)</param>
-	/// <param name="treeStyle">Tree texture index</param>
-	/// <param name="floorY"></param>
-	/// <param name="topTextureFrameWidth">Tree top width</param>
-	/// <param name="topTextureFrameHeight">Tree top height</param>
-	/// <returns></returns>
-	public static bool GetTreeFoliageData(int x, int y, int xoffset, ref int treeFrame, ref int treeStyle, out int floorY, out int topTextureFrameWidth, out int topTextureFrameHeight)
-	{
-		Tile tile = Main.tile[x, y];
-
-		floorY = 0;
-		topTextureFrameHeight = 0;
-		topTextureFrameWidth = 0;
-
-		if (!TryGetByTile(tile.TileType, out ModCustomTree tree))
-			return false;
-
-		treeStyle = tree.TreeStyle;
-		return tree.GetTreeFoliageData(x, y, xoffset, ref treeFrame, out floorY, out topTextureFrameWidth, out topTextureFrameHeight);
-	}
-
-	/// <summary>
-	/// Gets tree style (texture index)
-	/// </summary>
-	/// <param name="x">World X coordinate</param>
-	/// <param name="y">World Y coordinate</param>
-	/// <param name="style"></param>
-	public static void GetStyle(int x, int y, ref int style)
-	{
-		Tile tile = Main.tile[x, y];
-		if (!TryGetByTile(tile.TileType, out ModCustomTree tree))
-			return;
-
-		style = tree.GetStyle(x, y);
-	}
-
-	/// <summary>
-	/// Called when tree was shook
-	/// </summary>
-	/// <param name="x"></param>
-	/// <param name="y"></param>
-	/// <param name="createLeaves"></param>
-	/// <returns></returns>
-	public static bool Shake(int x, int y, ref bool createLeaves)
-	{
-		Tile tile = Main.tile[x, y];
-		if (!TryGetByTile(tile.TileType, out ModCustomTree tree))
-			return true;
-
-		return tree.Shake(x, y, ref createLeaves);
-	}
-
-	/// <summary>
-	/// Sets passStyle to tree leaf gore id
-	/// </summary>
-	/// <param name="x"></param>
-	/// <param name="topTile"></param>
-	/// <param name="t"></param>
-	/// <param name="treeHeight"></param>
-	/// <param name="treeFrame"></param>
-	/// <param name="passStyle"></param>
-	public static void GetTreeLeaf(int x, Tile topTile, Tile t, ref int treeHeight, ref int treeFrame, ref int passStyle)
-	{
-		if (!TryGetByTile(topTile.TileType, out ModCustomTree tree))
-			return;
-
-		tree.GetTreeLeaf(x, topTile, t, ref treeHeight, ref treeFrame, out passStyle);
-	}
-
-	/// <summary>
-	/// Tries to grow tree by tile type
-	/// </summary>
-	/// <param name="tileType"></param>
-	/// <param name="x"></param>
-	/// <param name="y"></param>
-	/// <returns></returns>
-	public static bool Grow(ushort tileType, int x, int y)
-	{
-		if (!TryGetByTile(tileType, out ModCustomTree tree))
-			return false;
-
-		return tree.Grow(x, y);
-	}
-
-	/// <summary>
-	/// Tries to grow tree by sapling tile type
-	/// </summary>
-	/// <param name="tileType"></param>
-	/// <param name="x"></param>
-	/// <param name="y"></param>
-	/// <returns></returns>
-	public static bool GrowFromSapling(ushort tileType, int x, int y)
-	{
-		if (!TryGetBySapling(tileType, out ModCustomTree tree))
-			return false;
-
-		return tree.Grow(x, y);
-	}
-
-	/// <summary>
-	/// Tries to generate any ModCustomTree at given coordinates
-	/// </summary>
-	/// <param name="x"></param>
-	/// <param name="y"></param>
-	/// <returns></returns>
-	public static bool TryGenerate(int x, int y)
-	{
-		if (trees.Count == 0)
-			return false;
-
-		int tileType = Framing.GetTileSafely(x, y).TileType;
-
-		int weight = 0;
-		foreach (ModCustomTree tree in trees)
-			if (tree.ValidGroundType(tileType))
-				weight += Math.Abs(tree.GenerationWeight);
-
-		if (weight <= 0)
-			return false;
-
-		int chosenWeight = WorldGen.genRand.Next(weight);
-		foreach (ModCustomTree tree in trees) {
-			if (!tree.ValidGroundType(tileType))
-				continue;
-
-			chosenWeight -= Math.Abs(tree.GenerationWeight);
-			if (chosenWeight > 0)
-				continue;
-
-			if (WorldGen.genRand.Next(tree.GenerationChance) == 0) {
-				return tree.TryGenerate(x, y);
-			}
-
-			return false;
-		}
-		return false;
-	}
-
-	/// <summary>
-	/// Returns true if given acorn item type can grow on given ground
-	/// </summary>
-	/// <param name="acornItemType"></param>
-	/// <param name="groundTileType"></param>
-	/// <returns></returns>
-	public static bool CanGrowAcorn(int acornItemType, ushort groundTileType) {
-		if (!TryGetByAcorn(acornItemType, out ModCustomTree tree))
-			return false;
-
-		return tree.ValidGroundType(groundTileType);
-	}
-}
-
 public abstract class ModCustomTree : ModType, ILocalizedModType
 {
 	/// <summary>
@@ -497,7 +255,7 @@ public abstract class ModCustomTree : ModType, ILocalizedModType
 	/// </summary>
 	public virtual void TileFrame(int x, int y)
 	{
-		CustomTreeGen.CheckTree(x, y, GetTreeSettings());
+		CustomTreeGenerator.CheckTree(x, y, GetTreeSettings());
 	}
 
 	/// <summary>
@@ -508,7 +266,7 @@ public abstract class ModCustomTree : ModType, ILocalizedModType
 	/// <returns></returns>
 	public virtual bool Grow(int x, int y)
 	{
-		if (CustomTreeGen.GrowTree(x, y, GetTreeSettings()) && WorldGen.PlayerLOS(x, y))
+		if (CustomTreeGenerator.GrowTree(x, y, GetTreeSettings()) && WorldGen.PlayerLOS(x, y))
 		{
 			WorldGen.TreeGrowFXCheck(x, y);
 			return true;
@@ -867,1103 +625,4 @@ internal class CustomTreeDefaultLeaf : ModGore
 	{
 		CustomTreeDefaults.SetLeafStaticDefaults(Type);
 	}
-}
-
-public static class CustomTreeGen
-{
-	static bool prevLeftBranch = false;
-	static bool prevRightBranch = false;
-
-	/// <summary>
-	/// Tries to grow a tree at the given coordinates with the given settings, returns true on success
-	/// </summary>
-	public static bool GrowTree(int x, int y, CustomTreeGenerationSettings settings)
-	{
-		prevLeftBranch = false;
-		prevRightBranch = false;
-
-		int groundY = y;
-		while (TileID.Sets.TreeSapling[Main.tile[x, groundY].TileType] || TileID.Sets.CommonSapling[Main.tile[x, groundY].TileType]) {
-			groundY++;
-		}
-		if (Main.tile[x - 1, groundY - 1].LiquidAmount != 0 || Main.tile[x, groundY - 1].LiquidAmount != 0 || Main.tile[x + 1, groundY - 1].LiquidAmount != 0) {
-			return false;
-		}
-		Tile ground = Main.tile[x, groundY];
-		if (!ground.HasUnactuatedTile || ground.IsHalfBlock || ground.Slope != SlopeType.Solid) {
-			return false;
-		}
-		if (!settings.GroundTypeCheck(ground.TileType) || !settings.WallTypeCheck(Main.tile[x, groundY - 1].WallType)) {
-			return false;
-		}
-
-		Tile groundLeft = Main.tile[x - 1, groundY];
-		Tile groundRight = Main.tile[x + 1, groundY];
-		if (
-			   (!groundLeft.HasTile || !settings.GroundTypeCheck(groundLeft.TileType))
-			&& (!groundRight.HasTile || !settings.GroundTypeCheck(groundRight.TileType))) {
-			return false;
-		}
-		byte color = Main.tile[x, groundY].TileColor;
-		int maxTreeHeight = settings.MinHeight;
-		while (maxTreeHeight < settings.MaxHeight) {
-			if (!WorldGen.EmptyTileCheck(x - 2, x + 2, groundY - maxTreeHeight - settings.TopPaddingNeeded, groundY - 1, 20)) {
-				if (maxTreeHeight == settings.MinHeight)
-					return false;
-				break;
-			}
-			maxTreeHeight++;
-		}
-
-		int treeHeight = Math.Min(WorldGen.genRand.Next(settings.MinHeight, maxTreeHeight + 1), settings.MaxHeight);
-		int treeBottom = groundY - 1;
-		int treeTop = treeBottom - treeHeight;
-
-		for (int i = treeBottom; i >= treeTop; i--) {
-			if (i == treeBottom)
-				PlaceBottom(x, i, color, settings, treeHeight == 1);
-			else if (i > treeTop)
-				PlaceMiddle(x, i, color, settings);
-			else
-				PlaceTop(x, i, color, settings);
-		}
-
-		WorldGen.RangeFrame(x - 2, treeTop - 1, x + 2, treeBottom + 1);
-		if (Main.netMode == NetmodeID.Server) {
-			NetMessage.SendTileSquare(-1, x - 1, treeTop, 3, treeHeight, TileChangeType.None);
-		}
-
-		return true;
-	}
-
-	/// <summary>
-	/// Generates tree botoom part at the given coordinates
-	/// </summary>
-	/// <param name="x">X coordinate</param>
-	/// <param name="y">Y coordinate</param>
-	/// <param name="color">Tile color</param>
-	/// <param name="settings">Tree settings</param>
-	/// <param name="top">True only if thee height is 1, places <see cref="TreeTileType.TopWithRoots"/> instead of <see cref="TreeTileType.WithRoots"/></param>
-	public static void PlaceBottom(int x, int y, byte color, CustomTreeGenerationSettings settings, bool top)
-	{
-		Tile groundRight = Framing.GetTileSafely(x + 1, y + 1);
-		Tile groundLeft = Framing.GetTileSafely(x - 1, y + 1);
-
-		bool rootRight = !WorldGen.genRand.NextBool(settings.NoRootChance)
-			&& groundRight.HasUnactuatedTile
-			&& !groundRight.IsHalfBlock
-			&& groundRight.Slope == SlopeType.Solid;
-
-		bool rootLeft = !WorldGen.genRand.NextBool(settings.NoRootChance)
-			&& groundLeft.HasUnactuatedTile
-			&& !groundLeft.IsHalfBlock
-			&& groundLeft.Slope == SlopeType.Solid;
-
-		int style = WorldGen.genRand.Next(3);
-
-		if (rootRight)
-			Place(x + 1, y, new(style, TreeTileSide.Right, TreeTileType.Root), color, settings);
-		if (rootLeft)
-			Place(x - 1, y, new(style, TreeTileSide.Left, TreeTileType.Root), color, settings);
-
-		if (rootLeft || rootRight)
-			Place(x, y, new(style, GetSide(rootLeft, rootRight), top ? TreeTileType.TopWithRoots : TreeTileType.WithRoots), color, settings);
-		else
-			PlaceNormal(x, y, color, settings);
-	}
-
-	/// <summary>
-	/// Places middle tree part
-	/// </summary>
-	public static void PlaceMiddle(int x, int y, byte color, CustomTreeGenerationSettings settings)
-	{
-		bool branchRight = WorldGen.genRand.NextBool(settings.BranchChance);
-		bool branchLeft = WorldGen.genRand.NextBool(settings.BranchChance);
-
-		int style = WorldGen.genRand.Next(3);
-
-		if (prevLeftBranch && branchLeft)
-			branchLeft = false;
-		if (prevRightBranch && branchRight)
-			branchRight = false;
-
-		prevLeftBranch = branchLeft;
-		prevRightBranch = branchRight;
-
-		if (branchRight)
-			Place(x + 1, y,
-				new(style, TreeTileSide.Right,
-				WorldGen.genRand.NextBool(settings.NotLeafyBranchChance) ? TreeTileType.Branch : TreeTileType.LeafyBranch),
-				color, settings);
-		if (branchLeft)
-			Place(x - 1, y,
-				new(style, TreeTileSide.Left,
-				WorldGen.genRand.NextBool(settings.NotLeafyBranchChance) ? TreeTileType.Branch : TreeTileType.LeafyBranch),
-				color, settings);
-
-		if (branchRight || branchLeft)
-			Place(x, y, new(style, GetSide(branchLeft, branchRight), TreeTileType.WithBranches), color, settings);
-		else
-			PlaceNormal(x, y, color, settings);
-	}
-
-	/// <summary>
-	/// Places tree top
-	/// </summary>
-	public static void PlaceTop(int x, int y, byte color, CustomTreeGenerationSettings settings)
-	{
-		if (WorldGen.genRand.NextBool(settings.BrokenTopChance))
-			Place(x, y, new(TreeTileType.BrokenTop), color, settings);
-		else
-			Place(x, y, new(TreeTileType.LeafyTop), color, settings);
-	}
-
-	/// <summary>
-	/// Places straight tree tile
-	/// </summary>
-	public static void PlaceNormal(int x, int y, byte color, CustomTreeGenerationSettings settings)
-	{
-		int bark = 0;
-
-		if (WorldGen.genRand.NextBool(settings.LessBarkChance))
-			bark--;
-		if (WorldGen.genRand.NextBool(settings.MoreBarkChance))
-			bark++;
-
-		TreeTileSide side = WorldGen.genRand.NextBool() ? TreeTileSide.Left : TreeTileSide.Right;
-
-		if (bark == 0)
-			Place(x, y, new(TreeTileType.Normal), color, settings);
-		else if (bark < 0)
-			Place(x, y, new(side, TreeTileType.LessBark), color, settings);
-		else
-			Place(x, y, new(side, TreeTileType.MoreBark), color, settings);
-	}
-
-	/// <summary>
-	/// Places tree tile
-	/// </summary>
-	/// <param name="x"></param>
-	/// <param name="y"></param>
-	/// <param name="info"></param>
-	/// <param name="color">Tile color</param>
-	/// <param name="settings"></param>
-	/// <param name="triggerTileframe">True for triggering TileFrame afterwards</param>
-	public static void Place(int x, int y, TreeTileInfo info, byte color, CustomTreeGenerationSettings settings, bool triggerTileframe = false)
-	{
-		Tile t = Main.tile[x, y];
-
-		t.HasTile = true;
-		t.TileType = settings.TreeTileType;
-		info.ApplyToTile(t);
-		t.TileColor = color;
-
-		if (triggerTileframe) {
-			WorldGen.TileFrame(x - 1, y, false, false);
-			WorldGen.TileFrame(x + 1, y, false, false);
-			WorldGen.TileFrame(x, y - 1, false, false);
-			WorldGen.TileFrame(x, y + 1, false, false);
-		}
-	}
-
-	/// <summary>
-	/// Construct <see cref="TreeTileSide"/> from left and right
-	/// </summary>
-	public static TreeTileSide GetSide(bool left, bool right)
-	{
-		if (left && right || !left && !right)
-			return TreeTileSide.Center;
-		if (left)
-			return TreeTileSide.Left;
-		return TreeTileSide.Right;
-	}
-
-	/// <summary>
-	/// Deconstruct <see cref="TreeTileSide"/>
-	/// </summary>
-	public static void SetSide(TreeTileSide side, out bool left, out bool right)
-	{
-		left = right = false;
-
-		if (side == TreeTileSide.Center)
-			left = right = true;
-		else if (side == TreeTileSide.Left)
-			left = true;
-		else
-			right = true;
-	}
-
-	/// <summary>
-	/// Gets tree statistics at the given coordinates
-	/// </summary>
-	public static TreeStats GetTreeStats(int x, int y)
-	{
-		TreeStats stats = new();
-
-		stats.Top = new(x, y);
-		stats.Bottom = new(x, y);
-
-		foreach (PositionedTreeTile tile in EnumerateTreeTiles(x, y)) {
-			stats.TotalBlocks++;
-
-			switch (tile.Info.Type) {
-				case TreeTileType.Branch:
-					stats.TotalBranches++;
-					break;
-
-				case TreeTileType.LeafyBranch:
-					stats.TotalBranches++;
-					stats.LeafyBranches++;
-					break;
-
-				case TreeTileType.Root:
-					if (tile.Info.Side == TreeTileSide.Left)
-						stats.LeftRoot = true;
-					else
-						stats.RightRoot = true;
-					break;
-
-				case TreeTileType.BrokenTop:
-					stats.HasTop = true;
-					stats.BrokenTop = true;
-					break;
-
-				case TreeTileType.LeafyTop:
-					stats.HasTop = true;
-					break;
-			}
-			if (tile.Info.IsCenter) {
-				stats.Bottom.X = tile.Pos.X;
-				stats.Top.X = tile.Pos.X;
-
-				stats.Top.Y = Math.Min(tile.Pos.Y, stats.Top.Y);
-				stats.Bottom.Y = Math.Max(tile.Pos.Y, stats.Bottom.Y);
-			}
-		}
-		stats.GroundType = Framing.GetTileSafely(stats.Bottom.X, stats.Bottom.Y + 1).TileType;
-		return stats;
-	}
-
-	/// <summary>
-	/// Find and enumerate throug all tiles of this tree
-	/// </summary>
-	public static IEnumerable<PositionedTreeTile> EnumerateTreeTiles(int x, int y)
-	{
-		HashSet<Point> done = new();
-		Queue<Point> queue = new();
-
-		queue.Enqueue(new Point(x, y));
-
-		while (queue.Count > 0) {
-			Point p = queue.Dequeue();
-			if (done.Contains(p))
-				continue;
-
-			done.Add(p);
-
-			Tile t = Main.tile[p.X, p.Y];
-			if (!t.HasTile || !TileID.Sets.IsATreeTrunk[t.TileType])
-				continue;
-
-			TreeTileInfo info = TreeTileInfo.GetInfo(t);
-
-			yield return new(p, info);
-
-			bool left = false;
-			bool right = false;
-			bool up = true;
-			bool down = true;
-
-			switch (info.Type) {
-				case TreeTileType.WithBranches:
-					SetSide(info.Side, out left, out right);
-					break;
-
-				case TreeTileType.Branch:
-					up = down = false;
-					SetSide(info.Side, out right, out left);
-					break;
-
-				case TreeTileType.LeafyBranch:
-					up = down = false;
-					SetSide(info.Side, out right, out left);
-					break;
-
-				case TreeTileType.WithRoots:
-					down = false;
-					SetSide(info.Side, out left, out right);
-					break;
-
-				case TreeTileType.Root:
-					up = down = false;
-					SetSide(info.Side, out right, out left);
-					break;
-
-				case TreeTileType.BrokenTop:
-					up = false;
-					break;
-
-				case TreeTileType.LeafyTop:
-					up = false;
-					break;
-			}
-
-			if (up)
-				queue.Enqueue(new(p.X, p.Y - 1));
-			if (down)
-				queue.Enqueue(new(p.X, p.Y + 1));
-			if (left)
-				queue.Enqueue(new(p.X - 1, p.Y));
-			if (right)
-				queue.Enqueue(new(p.X + 1, p.Y));
-		}
-	}
-
-	/// <summary>
-	/// CustomTree variant of <see cref="WorldGen.CheckTree"/>
-	/// </summary>
-	/// <param name="x"></param>
-	/// <param name="y"></param>
-	/// <param name="settings">Tree settings</param>
-	/// <param name="breakTiles">True to validate and break tree tiles, for example if nothing is below vertical tile</param>
-	/// <param name="fixFrames">True to fix tile frames, for example change <see cref="TreeTileType.WithBranches"/> to <see cref="TreeTileType.Normal"/> when all branches are cut off from tile</param>
-	public static void CheckTree(int x, int y, CustomTreeGenerationSettings settings, bool breakTiles = true, bool fixFrames = true)
-	{
-		Tile t = Framing.GetTileSafely(x, y);
-		if (t.TileType != settings.TreeTileType)
-			return;
-
-		TreeTileInfo info = TreeTileInfo.GetInfo(x, y);
-
-		if (breakTiles) {
-			bool groundCheck = info.Type != TreeTileType.Branch && info.Type != TreeTileType.LeafyBranch;
-			if (groundCheck) {
-				Tile ground = Framing.GetTileSafely(x, y + 1);
-				if (!ground.HasTile || ground.TileType != settings.TreeTileType && !settings.GroundTypeCheck(ground.TileType)) {
-					WorldGen.KillTile(x, y);
-					return;
-				}
-			}
-
-			if (info.Type == TreeTileType.Branch || info.Type == TreeTileType.LeafyBranch || info.Type == TreeTileType.Root) {
-				int checkX = x;
-				if (info.Side == TreeTileSide.Left)
-					checkX++;
-				else if (info.Side == TreeTileSide.Right)
-					checkX--;
-
-				if (Framing.GetTileSafely(checkX, y).TileType != settings.TreeTileType) {
-					WorldGen.KillTile(x, y);
-					return;
-				}
-			}
-		}
-
-		if (fixFrames) {
-			bool top = Framing.GetTileSafely(x, y - 1).TileType == settings.TreeTileType;
-			bool left = Framing.GetTileSafely(x - 1, y).TileType == settings.TreeTileType;
-			bool right = Framing.GetTileSafely(x + 1, y).TileType == settings.TreeTileType;
-
-			TreeTileInfo leftInfo = TreeTileInfo.GetInfo(x - 1, y);
-			TreeTileInfo rightInfo = TreeTileInfo.GetInfo(x + 1, y);
-
-			bool leftBranch = left && leftInfo.Type == TreeTileType.Branch || leftInfo.Type == TreeTileType.LeafyBranch;
-			bool rightBranch = right && rightInfo.Type == TreeTileType.Branch || rightInfo.Type == TreeTileType.LeafyBranch;
-
-			bool leftRoot = left && leftInfo.Type == TreeTileType.Root;
-			bool rightRoot = right && rightInfo.Type == TreeTileType.Root;
-
-			TreeTileInfo newInfo = info;
-
-			if (newInfo.IsTop || newInfo.Type == TreeTileType.Branch || newInfo.Type == TreeTileType.LeafyBranch || newInfo.Type == TreeTileType.Root) {
-			}
-			else if (leftBranch || rightBranch) {
-				newInfo.Type = top ? TreeTileType.WithBranches : TreeTileType.TopWithBranches;
-				newInfo.Side = GetSide(leftBranch, rightBranch);
-			}
-			else if (leftRoot || rightRoot) {
-				newInfo.Type = top ? TreeTileType.WithRoots : TreeTileType.TopWithRoots;
-				newInfo.Side = GetSide(leftRoot, rightRoot);
-			}
-			else if (!top) {
-				newInfo.Type = TreeTileType.Top;
-			}
-			else if (
-				   newInfo.Type != TreeTileType.Normal
-				&& newInfo.Type != TreeTileType.LessBark
-				&& newInfo.Type != TreeTileType.MoreBark) {
-				int bark = 0;
-
-				if (WorldGen.genRand.NextBool(settings.LessBarkChance))
-					bark--;
-				if (WorldGen.genRand.NextBool(settings.MoreBarkChance))
-					bark++;
-
-				newInfo.Side = bark == 0 ? TreeTileSide.Center : WorldGen.genRand.NextBool() ? TreeTileSide.Left : TreeTileSide.Right;
-
-				if (bark == 0)
-					newInfo.Type = TreeTileType.Normal;
-				else if (bark < 0)
-					newInfo.Type = TreeTileType.LessBark;
-				else
-					newInfo.Type = TreeTileType.MoreBark;
-			}
-			if (info != newInfo) {
-				newInfo.ApplyToTile(x, y);
-
-				WorldGen.TileFrame(x - 1, y, false, false);
-				WorldGen.TileFrame(x + 1, y, false, false);
-				WorldGen.TileFrame(x, y - 1, false, false);
-				WorldGen.TileFrame(x, y + 1, false, false);
-			}
-		}
-	}
-
-	/// <summary>
-	/// Changes the given coordinates to point to tree bottom
-	/// </summary>
-	public static void GetTreeBottom(ref int x, ref int y)
-	{
-		TreeTileInfo info = TreeTileInfo.GetInfo(x, y);
-
-		if (info.Type == TreeTileType.Root || info.Type == TreeTileType.Branch || info.Type == TreeTileType.LeafyBranch) {
-			if (info.Side == TreeTileSide.Left)
-				x++;
-			else if (info.Side == TreeTileSide.Right)
-				x--;
-		}
-
-		Tile t = Framing.GetTileSafely(x, y);
-		while (y < Main.maxTilesY - 50 && (!t.HasTile || TileID.Sets.IsATreeTrunk[t.TileType] || t.TileType == 72)) {
-			y++;
-			t = Framing.GetTileSafely(x, y);
-		}
-	}
-}
-
-/// <summary>
-/// CustomTrees growing settings
-/// </summary>
-public struct CustomTreeGenerationSettings
-{
-	/// <summary>
-	/// Tree tile type
-	/// </summary>
-	public ushort TreeTileType;
-
-	/// <summary>
-	/// Check if ground tile type is valid to grow on
-	/// </summary>
-	public Func<int, bool> GroundTypeCheck;
-
-	/// <summary>
-	/// Check if behind wall type is valid to grow
-	/// </summary>
-	public Func<int, bool> WallTypeCheck;
-
-	/// <summary>
-	/// Minimum tree height
-	/// </summary>
-	public int MinHeight;
-
-	/// <summary>
-	/// Maximum tree height
-	/// </summary>
-	public int MaxHeight;
-
-	/// <summary>
-	/// Tree top leaves height in tiles (4 for vanilla)
-	/// </summary>
-	public int TopPaddingNeeded;
-
-	/// <summary>
-	/// 1 in X chance for not generating root
-	/// </summary>
-	public int NoRootChance;
-
-	/// <summary>
-	/// 1 in X chance for generating tile with less bark
-	/// </summary>
-	public int LessBarkChance;
-
-	/// <summary>
-	/// 1 in X chance for generating tile with more bark
-	/// </summary>
-	public int MoreBarkChance;
-
-	/// <summary>
-	/// 1 in X chance for generating a branch
-	/// </summary>
-	public int BranchChance;
-
-	/// <summary>
-	/// 1 in X chance that generated branch will be without leaves
-	/// </summary>
-	public int NotLeafyBranchChance;
-
-	/// <summary>
-	/// 1 in X chance that thee top will be broken
-	/// </summary>
-	public int BrokenTopChance;
-
-	/// <summary/>
-	public CustomTreeGenerationSettings(WorldGen.GrowTreeSettings vanillaSettings)
-	{
-		TreeTileType = vanillaSettings.TreeTileType;
-
-		GroundTypeCheck = (t) => vanillaSettings.GroundTest(t);
-		WallTypeCheck = (t) => vanillaSettings.WallTest(t);
-
-		MinHeight = vanillaSettings.TreeHeightMin;
-		MaxHeight = vanillaSettings.TreeHeightMax;
-
-		TopPaddingNeeded = vanillaSettings.TreeTopPaddingNeeded;
-
-		NoRootChance = 3;
-		LessBarkChance = 7;
-		MoreBarkChance = 7;
-		BranchChance = 4;
-		NotLeafyBranchChance = 3;
-		BrokenTopChance = 13;
-	}
-
-	/// <summary>
-	/// Settings for vanilla common tree
-	/// </summary>
-	public static CustomTreeGenerationSettings VanillaCommonTree = new() {
-		TreeTileType = TileID.Trees,
-
-		GroundTypeCheck = (t) => WorldGen.IsTileTypeFitForTree((ushort)t),
-		WallTypeCheck = WorldGen.DefaultTreeWallTest,
-
-		MinHeight = 5,
-		MaxHeight = 17,
-
-		TopPaddingNeeded = 4,
-		BranchChance = 4,
-		MoreBarkChance = 7,
-		LessBarkChance = 7,
-		NotLeafyBranchChance = 3,
-		BrokenTopChance = 13,
-		NoRootChance = 3,
-	};
-}
-
-/// <summary/>
-public enum TreeTileSide
-{
-	/// <summary>
-	/// If tile is left sided, or have branches or roots to the left
-	/// </summary>
-	Left,
-
-	/// <summary>
-	/// If tile have branches or roots at both sides, default value
-	/// </summary>
-	Center,
-
-	/// <summary>
-	/// If tile is right sided, or have branches or roots to the right
-	/// </summary>
-	Right
-}
-
-/// <summary>
-/// Tree tile with position and its info
-/// </summary>
-public record struct PositionedTreeTile(Point Pos, TreeTileInfo Info);
-
-/// <summary/>
-public enum TreeFoliageType
-{
-	/// <summary/>
-	Top,
-
-	/// <summary/>
-	LeftBranch,
-
-	/// <summary/>
-	RightBranch
-}
-
-/// <summary>
-/// Tree statistics
-/// </summary>
-public struct TreeStats
-{
-	/// <summary>
-	/// Total tree blocks
-	/// </summary>
-	public int TotalBlocks = 0;
-
-	/// <summary>
-	/// Total tree branches
-	/// </summary>
-	public int TotalBranches = 0;
-
-	/// <summary>
-	/// Total leafy branches
-	/// </summary>
-	public int LeafyBranches = 0;
-
-	/// <summary>
-	/// True if tree top exisis and not broken
-	/// </summary>
-	public bool HasTop = false;
-
-	/// <summary>
-	/// True if tree top exisis and broken
-	/// </summary>
-	public bool BrokenTop = false;
-
-	/// <summary>
-	/// True if thee have left root
-	/// </summary>
-	public bool LeftRoot = false;
-
-	/// <summary>
-	/// True if thee have right root
-	/// </summary>
-	public bool RightRoot = false;
-
-	/// <summary>
-	/// Tree top position
-	/// </summary>
-	public Point Top = new(0, int.MaxValue);
-
-	/// <summary>
-	/// Tree bottom position
-	/// </summary>
-	public Point Bottom = new(0, 0);
-
-	/// <summary>
-	/// Tree ground tile type
-	/// </summary>
-	public ushort GroundType = 0;
-
-	public TreeStats()
-	{
-	}
-
-	/// <summary/>
-	public override string ToString()
-	{
-		return $"T:{TotalBlocks} " +
-			$"B:{TotalBranches} " +
-			$"BL:{LeafyBranches} " +
-			$"{(HasTop ? BrokenTop ? "TB " : "TL " : "")}" +
-			$"{(LeftRoot ? "Rl " : "")}" +
-			$"{(RightRoot ? "Rr " : "")}" +
-			$"X:{Top.X} Y:{Top.Y}-{Bottom.Y} G:{GroundType}";
-	}
-}
-
-public struct TreeTileInfo
-{
-	/// <summary>
-	/// Tree tile style, range is [0, 2]
-	/// </summary>
-	public int Style;
-
-	/// <summary>
-	/// Tile side, Left, Center or Right<br/>
-	/// In case of any tile type with branches or roots Center means that there is branches or roots on both sides
-	/// </summary>
-	public TreeTileSide Side;
-
-	/// <summary>
-	/// Tree tile type
-	/// </summary>
-	public TreeTileType Type;
-
-	/// <summary>
-	/// True if tile is a leafy top or leafy branch
-	/// </summary>
-	public bool IsLeafy => Type == TreeTileType.LeafyBranch || Type == TreeTileType.LeafyTop;
-
-	/// <summary>
-	/// True if tile is not a leafy top or leafy branch
-	/// </summary>
-	public bool IsWoody => !IsLeafy;
-
-	/// <summary>
-	/// True if tile is a tree center tile (not root or branch)
-	/// </summary>
-	public bool IsCenter => Type != TreeTileType.Branch && Type != TreeTileType.LeafyBranch && Type != TreeTileType.Root;
-
-	/// <summary>
-	/// True if this tile can have branches
-	/// </summary>
-	public bool WithBranches => Type == TreeTileType.WithBranches || Type == TreeTileType.TopWithBranches;
-
-	/// <summary>
-	/// True if this tile can have roots
-	/// </summary>
-	public bool WithRoots => Type == TreeTileType.WithRoots || Type == TreeTileType.TopWithRoots;
-
-	/// <summary>
-	/// True if tile is any tree top
-	/// </summary>
-	public bool IsTop =>
-		   Type == TreeTileType.Top
-		|| Type == TreeTileType.TopWithBranches
-		|| Type == TreeTileType.TopWithRoots
-		|| Type == TreeTileType.BrokenTop
-		|| Type == TreeTileType.LeafyTop;
-
-	/// <summary/>
-	public TreeTileInfo(int style, TreeTileSide side, TreeTileType type)
-	{
-		Style = style;
-		Side = side;
-		Type = type;
-	}
-
-	/// <summary/>
-	public TreeTileInfo(TreeTileSide side, TreeTileType type)
-	{
-		Style = WorldGen.genRand.Next(3);
-		Side = side;
-		Type = type;
-	}
-
-	/// <summary/>
-	public TreeTileInfo(TreeTileType type)
-	{
-		Style = WorldGen.genRand.Next(3);
-		Side = TreeTileSide.Center;
-		Type = type;
-	}
-
-	/// <summary>
-	/// Gets the info about tree tile at given coordinates
-	/// </summary>
-	public static TreeTileInfo GetInfo(int x, int y) => GetInfo(Framing.GetTileSafely(x, y));
-
-	/// <summary>
-	/// Gets the info about given tree tile
-	/// </summary>
-	public static TreeTileInfo GetInfo(Tile t)
-	{
-		Point frame = new(t.TileFrameX, t.TileFrameY);
-
-		int frameSize = 22;
-
-		int style = (frame.Y & frameSize * 3) / frameSize % 3;
-		frame.Y /= frameSize * 3;
-		frame.X /= frameSize;
-
-		switch (frame.X) {
-			case 0:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Center, TreeTileType.Normal);
-					case 1:
-						return new(style, TreeTileSide.Left, TreeTileType.LessBark);
-					case 2:
-						return new(style, TreeTileSide.Right, TreeTileType.WithRoots);
-					case 3:
-						return new(style, TreeTileSide.Center, TreeTileType.BrokenTop);
-				}
-				break;
-
-			case 1:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Right, TreeTileType.LessBark);
-					case 1:
-						return new(style, TreeTileSide.Right, TreeTileType.MoreBark);
-					case 2:
-						return new(style, TreeTileSide.Right, TreeTileType.Root);
-					case 3:
-						return new(style, TreeTileSide.Center, TreeTileType.LeafyTop);
-				}
-				break;
-
-			case 2:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Right, TreeTileType.WithBranches);
-					case 1:
-						return new(style, TreeTileSide.Left, TreeTileType.MoreBark);
-					case 2:
-						return new(style, TreeTileSide.Left, TreeTileType.Root);
-					case 3:
-						return new(style, TreeTileSide.Left, TreeTileType.LeafyBranch);
-				}
-				break;
-
-			case 3:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Left, TreeTileType.Branch);
-					case 1:
-						return new(style, TreeTileSide.Right, TreeTileType.WithBranches);
-					case 2:
-						return new(style, TreeTileSide.Left, TreeTileType.WithRoots);
-					case 3:
-						return new(style, TreeTileSide.Right, TreeTileType.LeafyBranch);
-				}
-				break;
-
-			case 4:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Left, TreeTileType.WithBranches);
-					case 1:
-						return new(style, TreeTileSide.Right, TreeTileType.Branch);
-					case 2:
-						return new(style, TreeTileSide.Center, TreeTileType.WithRoots);
-				}
-				break;
-
-			case 5:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Center, TreeTileType.Top);
-					case 1:
-						return new(style, TreeTileSide.Center, TreeTileType.WithBranches);
-				}
-				break;
-
-			case 6:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Left, TreeTileType.TopWithBranches);
-					case 1:
-						return new(style, TreeTileSide.Right, TreeTileType.TopWithBranches);
-					case 2:
-						return new(style, TreeTileSide.Center, TreeTileType.TopWithBranches);
-				}
-				break;
-
-			case 7:
-				switch (frame.Y) {
-					case 0:
-						return new(style, TreeTileSide.Left, TreeTileType.TopWithRoots);
-					case 1:
-						return new(style, TreeTileSide.Right, TreeTileType.TopWithRoots);
-					case 2:
-						return new(style, TreeTileSide.Center, TreeTileType.TopWithRoots);
-				}
-				break;
-		}
-
-		return new(style, TreeTileSide.Center, TreeTileType.None);
-	}
-
-	/// <summary>
-	/// Applies this info to tree tile at given coordinates
-	/// </summary>
-	public void ApplyToTile(int x, int y) => ApplyToTile(Framing.GetTileSafely(x, y));
-
-	/// <summary>
-	/// Applies this info to this tree tile
-	/// </summary>
-	public void ApplyToTile(Tile t)
-	{
-		Point frame = default;
-
-		switch (Type) {
-			case TreeTileType.LessBark:
-				if (Side == TreeTileSide.Left)
-					frame = new(0, 3);
-				else
-					frame = new(1, 0);
-				break;
-
-			case TreeTileType.Branch:
-				if (Side == TreeTileSide.Left)
-					frame = new(3, 0);
-				else
-					frame = new(4, 3);
-				break;
-
-			case TreeTileType.BrokenTop:
-				frame = new(0, 9);
-				break;
-
-			case TreeTileType.MoreBark:
-				if (Side == TreeTileSide.Left)
-					frame = new(1, 3);
-				else
-					frame = new(2, 3);
-				break;
-
-			case TreeTileType.WithBranches:
-				if (Side == TreeTileSide.Left)
-					frame = new(4, 0);
-				else if (Side == TreeTileSide.Right)
-					frame = new(3, 3);
-				else
-					frame = new(5, 3);
-				break;
-
-			case TreeTileType.LeafyBranch:
-				if (Side == TreeTileSide.Left)
-					frame = new(2, 9);
-				else
-					frame = new(3, 9);
-				break;
-
-			case TreeTileType.WithRoots:
-				if (Side == TreeTileSide.Left)
-					frame = new(3, 6);
-				else if (Side == TreeTileSide.Right)
-					frame = new(0, 6);
-				else
-					frame = new(4, 6);
-				break;
-
-			case TreeTileType.Root:
-				if (Side == TreeTileSide.Left)
-					frame = new(2, 6);
-				else
-					frame = new(1, 6);
-				break;
-
-			case TreeTileType.Top:
-				frame = new(5, 0);
-				break;
-
-			case TreeTileType.TopWithBranches:
-				if (Side == TreeTileSide.Left)
-					frame = new(6, 0);
-				else if (Side == TreeTileSide.Right)
-					frame = new(6, 3);
-				else
-					frame = new(6, 6);
-				break;
-
-			case TreeTileType.TopWithRoots:
-				if (Side == TreeTileSide.Left)
-					frame = new(7, 0);
-				else if (Side == TreeTileSide.Right)
-					frame = new(7, 3);
-				else
-					frame = new(7, 6);
-				break;
-
-			case TreeTileType.LeafyTop:
-				frame = new(1, 9);
-				break;
-		}
-
-		frame.Y += Style;
-
-		int frameSize = 22;
-		t.TileFrameX = (short)(frame.X * frameSize);
-		t.TileFrameY = (short)(frame.Y * frameSize);
-	}
-
-	/// <summary/>
-	public override string ToString()
-	{
-		return $"{Side} {Type} ({Style})";
-	}
-
-	/// <summary/>
-	public override bool Equals(object obj)
-	{
-		return obj is TreeTileInfo info &&
-			   Style == info.Style &&
-			   Side == info.Side &&
-			   Type == info.Type;
-	}
-
-	/// <summary/>
-	public override int GetHashCode()
-	{
-		return HashCode.Combine(Style, Side, Type);
-	}
-
-	/// <summary/>
-	public static bool operator ==(TreeTileInfo a, TreeTileInfo b)
-	{
-		if (a.Type != b.Type)
-			return false;
-		if (a.Side != b.Side)
-			return false;
-		return a.Style == b.Style;
-	}
-
-	/// <summary/>
-	public static bool operator !=(TreeTileInfo a, TreeTileInfo b)
-	{
-		return a.Type != b.Type || a.Side != b.Side || a.Style != b.Style;
-	}
-}
-
-/// <summary/>
-public enum TreeTileType
-{
-	/// <summary>
-	/// Unspecified tree tile
-	/// </summary>
-	None,
-
-	/// <summary>
-	/// Straight tree tile
-	/// </summary>
-	Normal,
-
-	/// <summary>
-	/// Tile with less bark on <see cref="TreeTileInfo.Side"/> (Left/Right)
-	/// </summary>
-	LessBark,
-
-	/// <summary>
-	/// Tile with more bark on <see cref="TreeTileInfo.Side"/> (Left/Right)
-	/// </summary>
-	MoreBark,
-
-	/// <summary>
-	/// Tile with branches (All <see cref="TreeTileSide"/>s)
-	/// </summary>
-	WithBranches,
-
-	/// <summary>
-	/// Left or Right branch
-	/// </summary>
-	Branch,
-
-	/// <summary>
-	/// Left or Right branch with leaves
-	/// </summary>
-	LeafyBranch,
-
-	/// <summary>
-	/// Bottom tile with roots (All <see cref="TreeTileSide"/>s)
-	/// </summary>
-	WithRoots,
-
-	/// <summary>
-	/// Left or Right root
-	/// </summary>
-	Root,
-
-	/// <summary>
-	/// Cutted off top tile
-	/// </summary>
-	Top,
-
-	/// <summary>
-	/// Cutted off top with branches
-	/// </summary>
-	TopWithBranches,
-
-	/// <summary>
-	/// Cutted off top with roots
-	/// </summary>
-	TopWithRoots,
-
-	/// <summary>
-	/// Broken tree top
-	/// </summary>
-	BrokenTop,
-
-	/// <summary>
-	/// Tree top with leaves
-	/// </summary>
-	LeafyTop
 }
