+++ patches/tModLoader/Terraria/ModLoader/TreeFoliageType.cs
@@ -1,0 +1,14 @@
+﻿namespace Terraria.ModLoader;
+
+/// <summary/>
+public enum TreeFoliageType
+{
+	/// <summary/>
+	Top,
+
+	/// <summary/>
+	LeftBranch,
+
+	/// <summary/>
+	RightBranch
+}
